Create README.md
This directory contains files showing how I build a 2 layer Neural Network (NN) to do a planar data classification. The model gave an accuracy of 90%, you can explore the ipynb file to how it was done. The .py file shows the necessary utilities and packages needed to get this done. Do well to explore the .py file also.